requirements.txt
python-dotenv==1.1.1
python-olm==3.2.16
matrix-nio[e2e]==0.25.2
